Fix References(Role="xxx") intrinsic method.
References and ReferenceNames methods have optional Role argument. If it is
present, the MI_references/MI_referenceNames threw an exception, because the
role property was not present in given property list. But there is no
requirement to provide property list at all, the property list only limits nr.
of returned properties.

git-svn-id: https://svn.code.sf.net/p/pywbem/code/pywbem@613 84f79d66-d411-0410-a85a-a3183736c6e8