fixed TOCTOU error when validating peer's certificate
By TOCTOU it's meant time-of-check-time-of-use. Up to now, pywbem made two
connections for one request (applies just to ssl). The first one made the
verification (without the hostname check) and the second one was used for
request. No verification was done for the latter, which could be abused.

Peer's certificate is now validated when connecting over ssl. To prevent
man-in-the-middle attack, verification of hostname is also added. Peer's
hostname must match the commonName of its certificate. Or it must be contained
in subjectAltName (list of aliases). M2Crypto package is used for that purpose.
Thanks to it both security enhancements could be implemented quiete easily.
Downside is a new dependency added to pywbem. Verification can be skipped if
no_verification is set to False.

Certificate trust store can now be specified by user. Some default paths, valid
for several distributions, were added.

git-svn-id: https://svn.code.sf.net/p/pywbem/code/pywbem@627 84f79d66-d411-0410-a85a-a3183736c6e8