Fixed issue #33: WBEMConnection.last_raw_reply was not set to the current reply in case of parsing errors in the reply.
In addition, the last_* variables in WBEMConnection are now all four initialized in the constructor (previously, only two of them were initialized), but the initial value is now None, consistent with how they are set while waiting for a response.

In addition, updated the docstring for the last_* variables to state the initial value.

git-svn-id: https://svn.code.sf.net/p/pywbem/code/pywbem@680 84f79d66-d411-0410-a85a-a3183736c6e8