Remove functionality that causes mof_compiler to set toinstance flavor when tosubclass flavor is set on class qualifier compile.
This removes that functionality to force set the toinstance flavor when
the tosubclass flavor is set on creating qualifier declarations and
creating classes in the compiler.  However, it enables specific setting
of toinstance if it is defined in the flavor definition in the mof.
Adds tests specifically to confirm this behavior and modifies one test
that was written to assume that the toinstance qualifier got set.

Also fixes an issue in mof_compiler where somehow we had removed the import os which caused the mof_compiler to fail.

Adds to changes.rst