Incorporates a new test in test_mof_compiler to do a round-trip test with the schema mof where it first compiles the defined schema then writes the mof back and recompiles the mof it just wrote out. This is built into an entirely new test class that replaces the original schema compile class.
This pr also  changes adds a single variable to WBEMConnection to keep
a compile-ordered list of classes so that mof/other output order can
be maintained as the same order as the original compile. This is
required since this compiler does not have any symbol forward vision
so superclasses must always be compiled before subclasses.

Remove orig test