Changes to correct issue #263 where CIMClass.CIMMethod.CIMParameter tomof() output drops array indicator
The mof should output an array indicator "[]" when CIMParameter is_array
is set and with the array size if array_size is not None.  The code
correctly issues the one if array size is set but ignores the array
indicator output if only is_array is set.

Added tests for this fix and extended other tomof() tests. Note that a
first push had a problem in that the last two lines of some mof were
compressed into a single line because we removed the \n at the end
of the CIMMethod tomof(). However, this was to correct a problem where we were getting blank line before the }; We removed change so that there is
blank line sometimes before the final closing quote.  An issue has been
filed.

Rebase from master.