Add  warning message if validate_callback parameter on WBEMConnection is used.
Document not used warning in api documentation for cim_operations.WBEMConnection and cim_http.py. Make no mention of deprecation, just the warning the
the parameter is ignored with Python 3.  It seems logical to issue such a warning since if someone implemented the call back they really do have expectations of it being used.

Includes change to changes.rst for the issue.