Add documentation on test_client yaml extension & extend run_operations.py
Documented the extension to test_client yaml for pull operations in the
README.md file

Added option to run_cimoperation.py to separate out long-running
operations so we can do short test and long test. It is new option on
the cmd line and tests that are long_running have a special decorator.

We also cleaned up the pull tests and corrected some issues in all of
the tests.  Part of this was to add some common functions that compare
lists of instances and paths for equality.

Note that the long tests (-l option) now take several minutes whereas
short tests take about 30 seconds.