Add the individual pull operations to wbemcli
Adds each operation with its parameters and detailed help to wbemcli.
This includes all of the pull operations.

Added note to changes.rst.

Note that not all have been tested however since that is a manual
operation.

Change short names to match the standard for wbemcli (first letter
of every word in parameter name).