# This is a combination of 3 commits.
# The first commit's message is:
Add tests for listener

# This is the 2nd commit message:

minor changes

# This is the 3rd commit message:

Add a test class for listener and a test

This test creates a server, listener, filter, and subscription.

It then validates that the subscription and filter are
in the host and then deletes the subscription and filter and verifies
that they were deleted.  It does NOT create a indication receiver
or test for indications actually sent from server.

Adds note to changes.rst