Modify compiler to separate class properties from instance properties
When instances were created by the compiler, it got information from the
corresponding class but put the instance property values into the class
iteslf. Separated these so the instance properties go into the instance.

When instance is created in the instance in the compiler, we do not copy the qualifiers from the class any more. Also, the instance is created with only
the properties from the instance mof definition and not other properties
from the class.

Added test for create instances.  This is based on the class used in the
TestType.mof file using a pragma to copy the class into a new
test_instance.mof file.   The embedded instance properties in the
instance are commented out because there is still an issue with those
properties (see issue #147)

Remove code to build path and add tests

Removed the code in the compiler that attempted to build the path for
instances created since there is no requirement for the compiler to build
a complete instance so that key properties may not exist in the
instance definition.

Adds 3 tests for building instances to confirm at least some errors in
the mof are caught (property name not in class, duplicate property,
property value type does not match property type in class.

Fix issue documented in issue#406, expecting message attribute in
ValueError exception.