Add an optional string id to listener and filter
Adds an optional user defined string id to the listener
definition an another one to the filter definition.

The listener id is attached to the name of all destination and
listener objects if it exists.  The filter id is attached to the
name property of all filters after the listener id.

These are an informal mechanism for the user to identify instances
of filters and destination ojects that it has registered with the
WBEM server from the name property of the instances

Nothing is attached if these attributes are None.

Modify original submission per comments and add tests

Modified original submission to put ids within context instead of
prefixes, to separate them with ':', to test type and illegal
characters, and to add tests to run_cim_operations.py

Extended the whole suite of listener tests against pegasus and
created tests for all options of the filter and listener id based
name properties.