Implement the enumerateInstances iterator functions
Implement the methods to provide an enumerateInstances iterator that
will use either the pull operations or the EnumerateInstances operations
to enumerate instances.

This code makes the decision on the first call for each operation type whether the pull operations can be used and also returns an iterator instead of
a list of returned objects.  If the pull requests can be used for the
specific operation on the defined connection it is used for all subsequent
requests for the current WBEMConnection.

This works for all of the enumerate type methods including:
    EnumerateInstances  (IterEnumerateInstances)
    EnumerateInstanceNames (IterEnumerateInstanceNames)
    Associators
    AssociatorNames
    References
    ReferenceNames
    ExecQuery

The new methods effectively use the original method names prepended by
Iter. For example IterEnumerateInstances executes the code that tries
to first execute OpenEnumerateInstances and PullInstancesWithPath but if
those functions are not supported falls back to EnumerateInstances

Adds a parameter (use_pull_operations) to WBEMConnection to allow the
user to force use of either pull or non-pull operations for the Iter...
method. Since there is no requirement that all the pull operations be
implemented, it also implements an internal flag for each pull operation
type so that if, for example, the enumerateInstancesWithPaths is
implemented but OpenEnumerateInstancePaths is not implememented that
will be understood and pull will be executed for the functions that have
pull and the original request for those functions where the pull is not
implemented.

The response to the first request from the server determines whether that
pull sequence is implemented. If CIM_ERR_NOT_SUPPORTED is returned the
client assumes that the pull operations for that sequence are not
implemented.

The new functions are also implemented in wbemcli

This is currently a WIP as it only implements only IterEnumerateInstances
and IterEnumerateInstancPaths.

Adds a system level config variable for MaxObjectCount so that it can be
considered an options on the api. This config variable for MaxObjectCount
is used in cim_operations.py and wbemcli.

Add iter reference and associator operations

Implements tests in cim_operations.py.
Test largely complete at this point.

These new apis may use the original enumeration apis but to increase
compatibility between the pull and the original operations, they  set
host and namespace on iter results. This is because if the Iter... apis use
the pull operations, the result (for EnumerateInstancesWithPath and
EnumerteInstancePaths) include the host and namespace.  For the original
EnumerateInstances and EnumerateInstancePaths, the responses generally do
not include the host and namespace information. By adding this information
to the response instances/paths if it is missing, they make the responses
equal whichever operation was used by the Iter method.

Includes update to changes.rst