Replaces the DOM XML parser with SAX XML parser to improve efficiency of response parsing. This changes the parser for operation responses, indication requests and embedded instances. to use the sax parser. While this parser is not more cpu efficient there is an enormous improvement in memory usage.
This pr also adds some specific tests for the parser.

Add this enhancement to changes.rst

Modifys cim_operations debug to prettyprint. This code uses the minidom to
generate reply xml for output when debug is set. That should be the
only use of the minidom in the pywbem code.

A test was added for the tupletree primarily to compare the original dom
implementation with the sax implementation.

The xml_to_tupletree and dom_to_tupletree
functions were removed from tupletree but they
were moved to test_tupletree so that the
dom vs sax test can be retained.  This is both to be able to continue
using this test code but also because these functions are referenced
from code in the attic and this retains at least some definition of that
code.