Add cli argument option to wbemcli to load scripts.
The option -s / --script loads and executes a python file defined by the
value of the option.  This option may be repeated to load multiple
scripts or multiple scripts may be defined within a single
-s option.  Note that we call them scripts here and in the documentation.

Scripts are executed before the interactive console gets control.

There are some examples of a scripts defined in examples directory:

1. wbemcli-server.py - Creates a SERVER object representing a WBEMServer
for the connection.  This can be used to access pywbem WBEMServer
properties, methods, etc.

2. wbemcli-quit.py - Demo of terminating wbemcli from within a script.

3. wbemcli-displayargs.py - Demo of displaying input arguments.

4. test_wbemcli_script.sh - A shell script that demos scripts.

Change names of example files to use _ rather than -

Changes the names to match normal file naming.