Correct formating of argparse epilog
This fixes formatting of the argparse desc and epilog by adding a custom
formatter that has multiple  formatters.  I am not sure how far I trust
this fix in the future but found the proposal on stackoverflow and it is
also recommended solution. in a python issue. That is documented in the
code

In addition it:

- expanded the desc slightly.

- changes main and main call to get script name from args.

- fixes an issue with a missing parameter on the openQueryInstances
operation definition in wbemcli. This issue popped up during testing
with on pylint (this actually showed up after I had run a number of
pylints that did not show the error) but it announced that the rc
attribute for ClassName did not exist.

Conflicts:
	wbemcli