Fix issue with the sax parser and str xml
Python 3.4 does not accept str input to the sax parser.

However, rather than just setting this in xml_to_tuple we put the test
and conversion to utf8 selectively in several places to minimize the
effect.

The alternative would be to put the same code in a single place,
the xml_to_tuple function.