1. Error in test_client with tuple in associators response 2. add associators.yaml openenuminstances.yaml, openenuminstancepaths.yaml 3. Extend enumerateinstances.yaml 4. Modified the invokemethod tests in run_cim_operations.py to have only one operation per test and added a test that actually returns data from the server 5. Added comment line to cim_operations.py
Change add tests for some pull operations
Adds a number of mock tests and extends others including:

1. pullinstanceswithpath
2. openenumerateinstanceswithpath
3. referenceclasses
4. referenceinstances
5. associatorclasses
6. associatorinstances
7. invokemethod
8. pullinstancepaths
9. openenumerateinstancepaths

To do this we extended test_client.py and added a number of tests to
run_cim_operations.py (see below)

We broke the associator/references tests up into separate files for class
and instance tests.

Sadly, this still brings only up to about 64% coverage.

Note: Many of the tests are still of limited scope. We are trying to
add TODO comments to the yaml to indicate where more tests are needed.
Generally, we want the following for each test:
1. Simple test with default values
2. Test with all values set
3. At least one exception test.

Adds a new definition to the yaml (ignore_test:) that allows tests to be
built into the yaml but ignored during test runs.  These tests are
executed when the option to run a single test is executed but they are
ignored in the general test_client execution.

Extended the test_client.pyu pullresult to:
a. Fix issues of testing and displaying the result.context
b. test for namespace in the result.context.

Added a number of tests to run_cimoperations.py, primarily testing for the
PyWBEM_Person class/instances, etc. since that could be generalized for
other test environments.  If the PyWBEM_Person class does not exist
in the server, these tests are bypassed.  We added tests for: associatons,
associationnames, references, referencenames for both class and instances.
These form the basis for using the recorder to get data for mock tests since
they are simpler than the definition of most of the internal classes.

Add more tests to mock

Extended for associatornamesclass, associatornames instance,
referencenames class and reference names instance.

Also added the tests to cim_operations.py

Add more tests

Adds more tests and fixes issue with ordering failure

1. Extend sort to one more xml entity
2. Add tests for closeEnumeration, enumerateclasses,
enumerateclassnames, enumeratequalifiers, getclass, getqualifier,
openreferencepaths, openreferenceswithpath

Adds more tests

Adds tests for CreateInstance, DeleteInstance, ans openassociators