Added a check for unset URL target in recorder.
Analysis of the issue:

After an exception (e.g. ConnectionError), the recorder still gets control
to do its thing. The `http_request.target` may be `None` in such a case.
However, the code assumed it could be concatenated with a string, which
fails in that case with:

  File "/home/kschopmeyer/pywbem/githubpywbem/pywbem/pywbem/_recorder.py", line 411, in record
    http_request.target
TypeError: Can't convert 'NoneType' object to str implicitly

Solution:
- The `target` member will only be appended if not `None`.