Add an example scriptlets. (#599)
1. Scans a server and generates overall information on the namespaces
and the number of classes, instances throughout the server.  Note that
this
is tuned to OpenPegasus right now in that it includes a list of classes
where the enumerate operations actually kill the server.  These are
really internal and test classes within pegasus but it is clear that
pegasus does not protect itself in these cases since accessing from a
client can kill the server.

This is useful in openpegasus and could server as examples of writing
these scriptlets but not sure if we want to commit them or not.

Removed second example, association explorer for the moment.