More robustness and improved diagnostics in os_setup.py.
This change solves the `IndexError` raised in `os_setup.py` that was
reported in issue #556, and it makes the code more robust against
unexpected output of installer commands, and improves the diagnostics
in cases of unexpected output.

This change does not necessarily solve the root cause of the problem,
though.