Create optional detailed line by line report output showing code coverage.
Adds an optional environment variable for the make file to control
output of a detailed line by line coverage report in html   If this
variable is set (export PYWBEM_COVERAGE_REPORT = true) html output is
generated in the directory coverage_html that shows the files in the
pywbem directory with coloring for which lines are covered and missed.
This also creates temporary files with the suffix ,cover in
pywbem/pywbem directory.  All these files are destroyed by clobber.

We made this an optional variable because it is only needed occasionally
when working on coverage issues. There is no need to generate these
reports for every make test.

To generate the reports:
  1) set the env var (export PYWBEM_COVERAGE_REPORT=true
  2) run make test
or
  1) run make test PYWBEM_COVERAGE_REPORT=true