Prototype for a comand line client tool to interact with wbem servers.
Creates a prototype for a cmdline tool that meets the criteria we
have defined in the issue.  I concluded with a lot of help
from Andy that click is a logical platform for this in that
it has several contributions that directly help our effort including
the repl, config file support, etc. that are not easy to achieve
in argparse.

For now the tool is named wcli.

This first prototype contains:

1. The base code for the app and some support functions including a
   display function and primitive uri parser.

2. Commands for getclass and enumerateclasses based on the previous
   cli concepts of direct access to the apis.

3. New command groups for class and instance that morph the actions
   to group them around nouns representing the objects rather than
   the operations directly.

4. subcommands for the class operation, specifically:
   a. get
   b. enumerate
   c. names (which enumerates names) This one may be deleted
   d. references (gets references classes)
   e. associators(gets associator classes)

5. Subcommands for the instance group:
   a. get
   b. names - possibly delete
   c. enumerate
   d. references
   e. associators
   f. count - count of instances by class

6. Subcommands for qualifier group:
   a. get
   b. enumerate

7. Subcommands for server group:
   a. namespaces
   b. interop
   c. brand
   d. info
   e. profiles
   f. connection

This should be considered a prototype to determine direction and further
requirements to drive the final code.

At this point there are NO tests.
Disable spinner until we move to its own repository

Set cli  name to pywbemcli

Add a design document

Add sort to several  groups but only partly implemented