Fix issue 656, MaxObjectCount parameter of Pull operations
This parameter was inconsistent between the different pull...
operations. Made it positional for all 3 operations.  Also added
validation for the
parameters for these operations (Context and MaxObjectCount).

Added tests to mock tests for the new method validation and modified
run_cim_operations.py to match the changes in cim_operations.py. Note
that these changes should not impact clients since the keyword parameter
is still accepted and the tests we added should be caught by the server
or pywbem code in any case.