rebase subsequent to the pr else/finally fix
to cim_operations.  This keeps statistics on time to
execute for the cim operations.

This captures time/count information on all of the non-iter
operations and also the count of exceptions.

Also found error in _recorder and fixed that error, specifically
a misnamed variable ret that should have been named ret_dict

Fix issue causing travis failure

Fixed comment by Andy

Conflicts:
	pywbem/cim_operations.py

Create classes for keeping statistics and add statistics
to cim_operations.  This keeps statistics on time to
execute for the cim operations.

This captures time/count information on all of the non-iter
operations and also the count of exceptions.

Also found error in _recorder and fixed that error, specifically
a misnamed variable ret that should have been named ret_dict

Fix issue causing travis failure

Fixed comment by Andy

Added a new attribute in WBEM_Connection, last_operation_time.
This is set at the end of each operation if statistics are
enabled.  It is set from the return value of the stats.stop_timer(exc)
function.  This allows a user to capture the client time for each
operation rather than the averaging that is in stats.

Minor fix for bad edit

Clean up issue with the Exception as ... variable

Extended statistics to include request and reply size