Add recorder unit test to testsuite
Fix for issue #609 indent by 4 in yaml output

Add generated yaml to enumerateinstances.yaml and other
testclient yaml files.

* Added long createinstance to run_cimoperations.py to be able to create
new mock createinstance test.  Added the record result of this output
to createinstance.yaml but there are issues so marked ignore in the yaml

* Added new property for timedelta to PyWBEM_AllTypes class. Used to
test for timedelta specifically

* Fixed run_cim_operations.py and test_recorder.py to open utf-8 aware
recorder output file. This was a problem with python 2.7 in that the
file open is utf-8 unaware so yaml writes fail if utf-8 in output. Fixes
issue #663

Fixed issue where the pull response pywbem_response in yaml was not
correctly representing eos, context.  Test of this in the new recorder unit test.

Insured that test_recorder test output file is in testsuite and ignored.
Added this test file to .gitignore