Add __init__.py to subdirs testsuite and examples to tell pylint to search for config file.
pylint searches up the hiearchy for the config file for any directory
that it considers a module (with __init__.py).  Add this to examples and
testsuite to force that search if pylint is used within an editor or if
it is run against a particular file or directory.

This appears to be the rulesi from pylint 1.8 documentation:

1. pylintrc in the current working directory
2. .pylintrc in the current working directory
3. If the current working directory is in a Python module, Pylint searches
up the hierarchy of Python modules until it finds a pylintrc file. This
allows you to specify coding standards on a module-by-module basis. Of
course, a directory is judged to be a Python module if it contains an
__init__.py file.
4. The file named by environment variable PYLINTRC

Sadly, flake8 does not have this complete a ruleset.  It is this directory
or home.