Clean up logging code and correct errors in testing
Fixed issue where the wbemcli banner was not available to test because
it is generated by python interactive by adding that call to the test
script.

Correct erroris in using all as component.This was causing a recursion
error

Add coveragerc to dev_setup and set minimal version to see if this
fixes issue with python 2.6 travis tests

Extend wbemcli unit tests and logging unit tests

Clean up logging defaults and move from config to logging

Add log_filename in _logging where it should be.