Correct bash execution call for windows in test_wbemcli.py so that test_wbemcli.py executes correctly in both windows and linux. Change test_logging.py to use the LogCapture package instead of trying to work with log files. Eliminates problem with windows and deleting log files.
Correct slashes on test_wbemcli.py file name for windows back to
linux slashes. They are created on windows but we are executing the
script in cygwin.

Uses wbemcli.bat file in windows to correctly execute wbemcli.
Without the bat file, the wbemcli call in test_wbemcli.py was
not finding the executable.

Fix wbemcli.bat file. This issue was identified in issue # 727 but
cannot finish test of this code without that fix