fixes #734 Fix issue with old version of coverage
This fixes issue where coveralls forced installation of old version of
coverage that causes problems with other tests because of deprecation
warnings it issues.  The problem was fixed after version 4.03 of
coverage

Had to directly change .travis.yml so it reinstall/updates coverage after
the installation in the setu.pyp and reinstallation as a coveralls
dependency with an old version of coverage.

We added the version to setup.py but that is mostly for documentation.

Added comment to changes.rst