fixes #734 Fix issue with old version of coverage
This fixes issue where coveralls forced installation of old version of
coverage that causes problems with other tests because of deprecation
warnings it issues.  The problem was fixed after version 4.03 of
coverage
We had to directly change .travis.yml to reinstall/update coverage after
the installation in the setup and reinstallation as a coveralls
dependency with an old version.

We added the version to setup.py but that is mostly for documentation.

Added comment to changes.rst

Conflicts:
	docs/changes.rst