Defines the _logging.py module component of pywbem logging and unit tests for this code.
This modules defines:

a. A single class with methods to construct named loggers to be used by
pywbem, in  particular operations logging and http logging. This class
includes methods to construct named loggers from either a string input
that defines the named loggers (useful for cli, etc. tools) or directly
from the parameters of the log.  Since the PywbemLogger class is to
be a singleton (there is not reason for separate instances) we chose
to use just class mathods to force it to be a singleton. This class
provides the methods to create named loggers from sets of parameters
including (destination, log level, and log detail level). It creates
a dictionary entry for each created logger but that is just for display
purposes as the actual information is in the created named loggers.

b. A unit test for this class.

c. The names of the named loggers that will be supported by pywbem
CIMOperations, in particular pywbem.ops and pywbem.http.

There is no changes.rst documentation because this is just part of the
total commits for issue #691.

The remaining changes to use this logger will be another part of this issue fix and will include changes to cim_operations.py and recorder.py.

Incorporateis Andy's comments

Incorporated all comments except for:

    The refs to the lists of allowed strings (I have a question about
those.

    The issue about get_logger in general

    Should we be doing create_loggers or configure loggers

fix comments by A. Maier

1. Use of autoclass for loggers

2. Clean up _logging to fix comments and links.  Add test to get_logger
so a pywbemlogger is created if it does not exist.

3.Add tests to test_logger

Fix issue with refs to variables

Add experimental

Add experimental to changes.rst