pywbem_os_setup.sh script for installing prereq OS packages
This change adds a bash script `pywbem_os_setup.sh` which installs
prerequisite OS packages for a number of Linux distributions.

The script basically does the same as the `install_os` command of
`setup.py`, but it is simpler in that it does not enforce the minimum
package versions, and it does not build swig if it cannot be installed.
The script is standalone (nothing else but the script is needed to
run it).

The RTD documentation was updated to explain how to download and run
the script. The RTD documentation uses the :download: role of Sphinx
which hosts a copy of the script along with the documentation on RTD.
This ensures that each version of the documentation references the
matching version of the script.
However, we only have two versions of the documentation on RTD (the
latest version released to Pypi, and the latest master).

The README file (which is shown both on the Github project page
and on the Pypi download page) was updated to explain how to
download and run the script. Because these sites perform only
the RST processing but no Sphinx processing on the README file,
the approach with linking the file with the :download: role is
not possible here. The script is linked from a tagged branch
of the git repo on Github. For this reason, the README file
now specifies the package version in a substition variable
which needs to be manually maintained.

The instructions for releasing a version to Pypi have been updated
to include a step where the version in the README file is being
updated.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>