Fixed TypeError: Error when calling the metaclass bases.
Details:

The `PywbemLoggers` class was defined by inheriting from
`six.with_metaclass(MetaPywbemLoggers)` and it also defined
the `__metaclass__` attribute.

Version 0.11.0 of the `six` package does not deal with this anymore
and raises:

  TypeError: Error when calling the metaclass bases
  metaclass conflict: the metaclass of a derived class must be a
    (non-strict) subclass of the metaclasses of all its bases

This change removes setting the `__metaclass__` attribute, which
is not needed anyway when using `six.with_metaclass()`, and which
is (or will at some point?) no longer supported in Python 3.

See also discussion in six issue #210:
https://github.com/benjaminp/six/issues/210

Signed-off-by: Andreas Maier <maiera@de.ibm.com>