Fix issues with iter methods
1. Did not include the IncludeQualifiers parameter in two calls to the
underlying functions from the Itermethods.

2. Used the name=value call defintion for the calls from
IterQueryInstances to ExecQuery and OpenQueryInstances.  While not
incorrect, these are required parameters so this is confusing in the
code and also means that the mock returns Name=value rather than value
for the test on method input.  I propose that we change these to remove
the Name= component.